# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Nibras - Intelligent Assistant for New Teachers
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudentLevel =
  (typeof StudentLevel)[keyof typeof StudentLevel];

export const StudentLevel = {
  مبتدئ: "مبتدئ",
  متوسط: "متوسط",
  متقدم: "متقدم",
  مختلط: "مختلط",
} as const;

export interface LessonRequest {
  /** Name of the teacher */
  teacherName: string;
  /** Educational stage (Primary, Middle, High) */
  stage: string;
  /** Grade level */
  grade: string;
  /** Subject name */
  subject: string;
  /** Title of the lesson */
  lessonTitle: string;
  /** Lesson duration in minutes (default 45) */
  duration?: number;
  /** Expected learning outcomes for the lesson */
  learningOutcomes?: string;
  /** Overall level of students in the class */
  studentLevel?: StudentLevel;
  /** Nature or description of the class */
  classNature?: string;
  /** Special learning needs or accommodations (optional) */
  learningNeeds?: string;
}

export interface TimeDistribution {
  /** Minutes for introduction */
  introduction: number;
  /** Minutes for explanation */
  explanation: number;
  /** Minutes for activities */
  activity: number;
  /** Minutes for assessment */
  assessment: number;
}

export type ActivityType = (typeof ActivityType)[keyof typeof ActivityType];

export const ActivityType = {
  individual: "individual",
  group: "group",
  discussion: "discussion",
  practical: "practical",
  digital: "digital",
} as const;

export interface Activity {
  name: string;
  description: string;
  duration: number;
  type: ActivityType;
}

export interface DifferentiatedInstructions {
  /** Instructions for advanced students */
  advanced: string;
  /** Instructions for average students */
  average: string;
  /** Instructions for students needing support */
  needsSupport: string;
}

export interface LessonPlan {
  title: string;
  stage: string;
  grade: string;
  subject: string;
  duration: number;
  /** Behavioral/procedural objectives */
  objectives: string[];
  timeDistribution: TimeDistribution;
  /** Active learning strategies */
  strategies: string[];
  introduction: string;
  mainContent: string;
  activities: Activity[];
  assessment: string;
  closure: string;
  materials: string[];
  differentiatedInstructions: DifferentiatedInstructions;
  homework?: string;
  madrasatiHomework?: string;
}

export interface LessonResponse {
  lessonPlan: LessonPlan;
  /** Sources used for RAG context */
  ragSources: string[];
  /** ISO timestamp of generation */
  generatedAt: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatRequest {
  /** User message */
  message: string;
  lessonContext?: LessonPlan;
  history?: ChatMessage[];
}

export interface LessonPlanPatch {
  title?: string;
  objectives?: string[];
  introduction?: string;
  mainContent?: string;
  activities?: Activity[];
  assessment?: string;
  closure?: string;
  materials?: string[];
  strategies?: string[];
  differentiatedInstructions?: DifferentiatedInstructions;
  homework?: string;
  madrasatiHomework?: string;
}

export interface ChatResponse {
  /** AI response */
  reply: string;
  /** RAG sources used to enrich the response */
  ragSources?: string[];
  /** Partial lesson plan patch to apply when Nuha modifies the plan */
  lessonPatch?: LessonPlanPatch;
  timestamp: string;
}

export interface RagRequest {
  subject: string;
  lessonTitle: string;
  stage?: string;
  grade?: string;
}

export interface RagResponse {
  sources: string[];
  context: string;
  retrievedAt: string;
}

export interface Subject {
  name: string;
  lessons: string[];
}

export interface Grade {
  name: string;
  subjects: Subject[];
}

export interface Stage {
  name: string;
  grades: Grade[];
}

export interface CurriculumOptions {
  stages: Stage[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}